feat: move agent state into AgentParticipant
